Combining Atlas code with old gazebo_plugins
